requirements.txt
# RHC Core Pipeline — Python Dependencies
# Install with: pip install -r requirements.txt
# Requires Python 3.12+

numpy>=1.26
pandas>=2.0
scipy>=1.12
neurokit2>=0.2.10  # 0.2.10+ fixes np.trapz removal (NumPy 2.0) and pandas CoW warnings
matplotlib>=3.8
PyYAML>=6.0
h5py>=3.10
openpyxl>=3.1
